final version

+++ Server/src/main/java/ua/woochat/server/model/ConfigServer.java
@@ -1,7 +1,6 @@
 package ua.woochat.server.model;
 
 import org.apache.log4j.Logger;
-
 import java.io.*;
 import java.util.Properties;
 
@@ -12,6 +11,7 @@ public class ConfigServer {
     private static ConfigServer configServer;
     private static Properties properties;
     private static final Logger logger = Logger.getLogger(ConfigServer.class);
+    private static final String EXTERNAL_PROPERTIES = "serverExtracted.properties";
 
 
     private ConfigServer() {
@@ -26,64 +26,66 @@ public class ConfigServer {
         return configServer;
     }
 
-    public static void loadServerConfig() {
+    private static void loadServerConfig() {
         try {
-            properties.load(ConfigServer.class.getClassLoader().getResourceAsStream("server.properties"));
-            File file = new File("serverExtracted.properties");
-            logger.debug("Server properties was extracted from jar file.");
-            logger.debug("Path to serverExtracted.properties: " + System.getProperty("user.dir") + File.separator + ("serverExtracted.properties"));
-            properties.store(new FileOutputStream(file), null);
+            File file = new File(EXTERNAL_PROPERTIES);
+            if (file.exists()) {
+                properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
+            } else {
+                properties.load(ConfigServer.class.getClassLoader().getResourceAsStream("server.properties"));
+                logger.debug("Server properties was extracted from jar file to: " + System.getProperty("user.dir") + File.separator + (EXTERNAL_PROPERTIES));
+                properties.store(new FileOutputStream(file), null);
+            }
         } catch (IOException e) {
-            e.printStackTrace();
+            logger.error("File not found exceptions ", e);
         }
     }
 
     public static String getRootAdmin() {
         try {
-            properties.load(new FileInputStream("serverExtracted.properties"));
+            properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
         } catch (IOException e) {
-            logger.error("IOException error " + e);
+            logger.error("IOException error ", e);
         }
         return properties.getProperty("rootAdmin");
     }
 
     public static String getPortConnection() {
         try {
-            properties.load(new FileInputStream("serverExtracted.properties"));
+            properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
         } catch (IOException e) {
-            logger.error("IOException error " + e);
+            logger.error("IOException error ", e);
         }
         return (properties.getProperty("portConnection"));
     }
 
     public static String getPortChatting() {
         try {
-            properties.load(new FileInputStream("serverExtracted.properties"));
+            properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
         } catch (IOException e) {
-            logger.error("IOException error " + e);
+            logger.error("IOException error ", e);
         }
         return (properties.getProperty("portChatting"));
     }
 
     public static String getTimeOut() {
         try {
-            properties.load(new FileInputStream("serverExtracted.properties"));
-            //properties.load(ConfigServer.class.getClassLoader().getResourceAsStream("serverExtracted.properties"));
+            properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
         } catch (IOException e) {
-            logger.error("IOException error " + e);
+            logger.error("IOException error ", e);
         }
         return (properties.getProperty("timeout"));
     }
 
     public static void setConfig(String type, String value) {
         try {
-            properties.load(new FileInputStream("serverExtracted.properties"));
+            properties.load(new FileInputStream(EXTERNAL_PROPERTIES));
             properties.setProperty(type, value);
-            File file = new File(System.getProperty("user.dir") + File.separator + "serverExtracted.properties");
+            File file = new File(System.getProperty("user.dir") + File.separator + EXTERNAL_PROPERTIES);
             logger.debug("Path to server.properties: " + ConfigServer.class.getClassLoader().getResourceAsStream("server.properties").toString());
             properties.store(new FileOutputStream(file), null);
         } catch (IOException e) {
-            e.printStackTrace();
+            logger.error("IOException error ", e);
         }
     }
 }
